+++ expression/exceptions/CheckedDivide.java
@@ -1,4 +1,5 @@
 package expression.exceptions;
+
 import expression.*;
 
 public class CheckedDivide extends Divide {
@@ -6,9 +7,9 @@ public class CheckedDivide extends Divide {
     public CheckedDivide(GlobalExpression first, GlobalExpression second) {
         super(first, second);
     }
-    
+
     @Override
-    public int calculate(int x, int y) {
+    public double calculate(double x, double y) {
         CalculationChecker.checkDivide(x, y);
         return x / y;
     }
