+++ expression/exceptions/CheckedMultiply.java
@@ -1,4 +1,5 @@
 package expression.exceptions;
+
 import expression.*;
 
 public class CheckedMultiply extends Multiply {
@@ -8,7 +9,7 @@ public class CheckedMultiply extends Multiply {
     }
 
     @Override
-    public int calculate(int x, int y) {
+    public double calculate(double x, double y) {
         CalculationChecker.checkMultiply(x, y);
         return x * y;
     }
