+++ expression/exceptions/CheckedNegate.java
@@ -1,4 +1,5 @@
 package expression.exceptions;
+
 import expression.*;
 
 public class CheckedNegate extends Negative {
@@ -6,9 +7,9 @@ public class CheckedNegate extends Negative {
     public CheckedNegate(GlobalExpression first) {
         super(first);
     }
-    
+
     @Override
-    public int calculate(int x) {
+    public double calculate(double x) {
         CalculationChecker.checkNegate(x);
         return -x;
     }
